replaced typeResolver with new implementation
