+++ logic/LogicCache.java
@@ -29,23 +29,18 @@ import java.util.Set;
 
 public class LogicCache {
 
-    private CommonCache<Conjunction, Map<Reference, Set<Label>>> typeHinterCache;
     private CommonCache<Traversal, Map<Reference, Set<Label>>> typeResolverCache;
     private CommonCache<String, Rule> ruleCache;
 
     public LogicCache() {
-        this.typeHinterCache = new CommonCache<>();
         this.ruleCache = new CommonCache<>();
         this.typeResolverCache = new CommonCache<>();
     }
 
     public LogicCache(int size, int timeOutMinutes) {
-        this.typeHinterCache = new CommonCache<>(size, timeOutMinutes);
         this.ruleCache = new CommonCache<>(size, timeOutMinutes);
         this.typeResolverCache = new CommonCache<>(size, timeOutMinutes);
     }
-
-    public CommonCache<Conjunction, Map<Reference, Set<Label>>> resolver() { return typeHinterCache; }
 
     public CommonCache<Traversal, Map<Reference, Set<Label>>> resolverTraversal() { return typeResolverCache; }
