+++ logic/LogicManager.java
@@ -27,7 +27,6 @@ import grakn.core.graph.GraphManager;
 import grakn.core.graph.structure.RuleStructure;
 import grakn.core.graph.util.Encoding;
 import grakn.core.logic.tool.TypeResolver;
-import grakn.core.logic.tool.TypeResolverTraversal;
 import grakn.core.traversal.TraversalEngine;
 import graql.lang.pattern.Conjunctable;
 import graql.lang.pattern.Conjunction;
@@ -49,16 +48,14 @@ public class LogicManager {
 
     private final ConceptManager conceptMgr;
     private final GraphManager graphMgr;
-//    private final TypeResolver typeResolver;
-    private final TypeResolverTraversal typeResolverTraversal;
+    private final TypeResolver typeResolver;
     private LogicCache logicCache;
 
     public LogicManager(GraphManager graphMgr, ConceptManager conceptMgr, TraversalEngine traversalEng, LogicCache logicCache) {
         this.graphMgr = graphMgr;
         this.conceptMgr = conceptMgr;
         this.logicCache = logicCache;
-//        this.typeResolver = new TypeResolver(conceptMgr, traversalEng, logicCache);
-        this.typeResolverTraversal = new TypeResolverTraversal(conceptMgr, traversalEng, logicCache);
+        this.typeResolver = new TypeResolver(conceptMgr, traversalEng, logicCache);
     }
 
     public Rule putRule(String label, Conjunction<? extends Pattern> when, ThingVariable<?> then) {
@@ -105,12 +102,8 @@ public class LogicManager {
                 .forEach(structure -> getRule(structure.label()).validateCycles());
     }
 
-//    public TypeResolver typeResolver() {
-//        return typeResolver;
-//    }
-
-    public TypeResolverTraversal typeResolver() {
-        return typeResolverTraversal;
+    public TypeResolver typeResolver() {
+        return typeResolver;
     }
 
     static void validateRuleStructureLabels(ConceptManager conceptMgr, RuleStructure ruleStructure) {
