+++ reasoner/Reasoner.java
@@ -65,7 +65,6 @@ public class Reasoner {
 
     public ResourceIterator<ConceptMap> execute(Disjunction disjunction, boolean isParallel) {
         Set<Conjunction> conjunctions = disjunction.conjunctions().stream()
-//                .map(conjunction -> logicMgr.typeResolver().resolveLabels(conjunction))
                 .map(conjunction -> logicMgr.typeResolver().resolveVariables(conjunction))
                 .collect(Collectors.toSet());
         // TODO enable: conjunction = logicMgr.typeResolver().resolveVariablesExhaustive(conjunction);
